Deleted epics.py, and redistributed its contents, to allow pyepics to work.
