+++ pytac/__init__.py
@@ -12,9 +12,9 @@ LIVE = 'live'
 # Default argument flag.
 DEFAULT = 'default'
 
-from . import data_source, element, epics, exceptions, lattice, load_csv, units, utils  # noqa: E402
+from . import data_source, element, exceptions, lattice, load_csv, units, utils  # noqa: E402
 """Error 402 is suppressed as we cannot import these modules at the top of the
 file as the strings above must be set first or the imports will fail.
 """
-__all__ = ["data_source", "element", "epics", "exceptions", "lattice",
-           "load_csv", "units", "utils"]
+__all__ = ["data_source", "element", "exceptions", "lattice", "load_csv",
+           "units", "utils"]
