+++ test/conftest.py
@@ -3,11 +3,11 @@ import mock
 import pytest
 import pytac
 from pytac import load_csv
-from pytac.element import Element
-from pytac.lattice import Lattice
+from pytac.element import Element, EpicsElement
+from pytac.lattice import Lattice, EpicsLattice
 from pytac.data_source import DataSourceManager, DeviceDataSource
 from pytac.units import PolyUnitConv
-from pytac.epics import EpicsLattice, EpicsElement, EpicsDevice
+from pytac.device import EpicsDevice
 from constants import DUMMY_VALUE_1, DUMMY_VALUE_2, RB_PV, SP_PV, LATTICE_NAME, CURRENT_DIR, DUMMY_ARRAY
